+++ rethinkdb-net/QueryTerm/ExpressionUtils.cs
@@ -2,6 +2,8 @@ using RethinkDb.Spec;
 using System;
 using System.Linq;
 using System.Linq.Expressions;
+using System.Reflection;
+using RethinkDb.Expressions;
 
 namespace RethinkDb.QueryTerm
 {
@@ -18,38 +20,8 @@ namespace RethinkDb.QueryTerm
 
         public static Term CreateFunctionTerm<TParameter1, TReturn>(IDatumConverterFactory datumConverterFactory, Expression<Func<TParameter1, TReturn>> expression)
         {
-            var funcTerm = new Term() {
-                type = Term.TermType.FUNC
-            };
-
-            var parametersTerm = new Term() {
-                type = Term.TermType.MAKE_ARRAY,
-            };
-            parametersTerm.args.Add(new Term() {
-                type = Term.TermType.DATUM,
-                datum = new Datum() {
-                    type = Datum.DatumType.R_NUM,
-                    r_num = 2
-                }
-            });
-            funcTerm.args.Add(parametersTerm);
-
-            var body = expression.Body;
-            if (body.NodeType == ExpressionType.MemberInit)
-            {
-                var memberInit = (MemberInitExpression)body;
-                if (!memberInit.Type.Equals(typeof(TReturn)))
-                    throw new InvalidOperationException("Only expression types matching the table type are supported");
-                else if (memberInit.NewExpression.Arguments.Count != 0)
-                    throw new NotSupportedException("Constructors will not work here, only field member initialization");
-                funcTerm.args.Add(MapMemberInitToTerm<TParameter1, TReturn>(datumConverterFactory, memberInit));
-            }
-            else
-            {
-                funcTerm.args.Add(ExpressionUtils.MapExpressionToTerm<TParameter1>(datumConverterFactory, expression.Body));
-            }
-
-            return funcTerm;
+            var converter = new SingleParameterLambda<TParameter1, TReturn>(datumConverterFactory);
+            return converter.CreateFunctionTerm(expression);
         }
 
         public static Term CreateFunctionTerm<TParameter1, TParameter2, TReturn>(IDatumConverterFactory datumConverterFactory, Expression<Func<TParameter1, TParameter2, TReturn>> expression)
@@ -93,101 +65,6 @@ namespace RethinkDb.QueryTerm
             }
 
             return funcTerm;
-        }
-
-        #endregion
-        #region Single Parameter Mappings
-
-        private static Term MapMemberInitToTerm<TParameter1, TReturn>(IDatumConverterFactory datumConverterFactory, MemberInitExpression memberInit)
-        {
-            var makeObjTerm = new Term() {
-                type = Term.TermType.MAKE_OBJ,
-            };
-
-            foreach (var binding in memberInit.Bindings)
-            {
-                switch (binding.BindingType)
-                {
-                    case MemberBindingType.Assignment:
-                        makeObjTerm.optargs.Add(MapMemberAssignmentToMakeObjArg<TParameter1, TReturn>(datumConverterFactory, (MemberAssignment)binding));
-                        break;
-                    case MemberBindingType.ListBinding:
-                    case MemberBindingType.MemberBinding:
-                        throw new NotSupportedException("Binding type not currently supported");
-                }
-            }
-
-            return makeObjTerm;
-        }
-
-        private static Term.AssocPair MapMemberAssignmentToMakeObjArg<TParameter1, TReturn>(IDatumConverterFactory datumConverterFactory, MemberAssignment memberAssignment)
-        {
-            var retval = new Term.AssocPair();
-
-            var datumConverter = datumConverterFactory.Get<TReturn>();
-            var fieldConverter = datumConverter as IObjectDatumConverter;
-            if (fieldConverter == null)
-                throw new NotSupportedException("Cannot map member assignments into ReQL without implementing IObjectDatumConverter");
-
-            retval.key = fieldConverter.GetDatumFieldName(memberAssignment.Member);
-            retval.val = ExpressionUtils.MapExpressionToTerm<TParameter1>(datumConverterFactory, memberAssignment.Expression);
-
-            return retval;
-        }
-
-        private static Term MapExpressionToTerm<TParameter1>(IDatumConverterFactory datumConverterFactory, Expression expr)
-        {
-            Func<Expression, Term> recursiveMap = (innerExpr) => MapExpressionToTerm<TParameter1>(datumConverterFactory, innerExpr);
-
-            switch (expr.NodeType)
-            {
-                case ExpressionType.MemberAccess:
-                {
-                    var memberExpr = (MemberExpression)expr;
-
-                    if (memberExpr.Expression.NodeType != ExpressionType.Parameter)
-                        throw new NotSupportedException("Unrecognized member access pattern");
-
-                    var getAttrTerm = new Term() {
-                        type = Term.TermType.GET_FIELD
-                    };
-
-                    getAttrTerm.args.Add(new Term() {
-                        type = Term.TermType.VAR,
-                        args = {
-                            new Term() {
-                                type = Term.TermType.DATUM,
-                                datum = new Datum() {
-                                    type = Datum.DatumType.R_NUM,
-                                    r_num = 2
-                                },
-                            }
-                        }
-                    });
-
-                    var datumConverter = datumConverterFactory.Get<TParameter1>();
-                    var fieldConverter = datumConverter as IObjectDatumConverter;
-                    if (fieldConverter == null)
-                        throw new NotSupportedException("Cannot map member access into ReQL without implementing IObjectDatumConverter");
-
-                    var datumFieldName = fieldConverter.GetDatumFieldName(memberExpr.Member);
-                    if (string.IsNullOrEmpty(datumFieldName))
-                        throw new NotSupportedException(String.Format("Member {0} on type {1} could not be mapped to a datum field", memberExpr.Member.Name, memberExpr.Type));
-
-                    getAttrTerm.args.Add(new Term() {
-                        type = Term.TermType.DATUM,
-                        datum = new Datum() {
-                            type = Datum.DatumType.R_STR,
-                            r_str = datumFieldName
-                        }
-                    });
-
-                    return getAttrTerm;
-                }
-
-                default:
-                    return SimpleMap(datumConverterFactory, recursiveMap, expr);
-            }
         }
 
         #endregion